refactor: HomeView 유틸 함수 분리 및 groupedTransactions 스토어 이동
